fusion-dhl/requirements.txt
numpy 
scipy
scikit-image
h5py
numpy-quaternion
matplotlib
tk
tensorboardx
torch
torchvision
utm
geopy
